% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.data.R
\docType{data}
\name{benchmark.data}
\alias{benchmark.data}
\title{Benchmark.data.}
\format{A list of dataframes with 3 columns and 200 samples, where the x and y
are generated from 2 multivariate distributions. The mean of the two groups vary,
to allow for different amount overlap for the groups.}
\usage{
benchmark.data
}
\description{
Datasets two groups, labels -1 and 1, with varying amounts of overlap.
}
\keyword{datasets}
