% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_mat_ops}
\alias{fast_mat_ops}
\alias{fMatProd}
\alias{fMatSolve}
\alias{fMatInv}
\title{Functions to use MKL to do the matrix calculations}
\usage{
fMatProd(x, y)

fMatSolve(x, y)

fMatInv(x)
}
\arguments{
\item{x, y}{matrices}
}
\value{
The result matrices
}
\description{
Functions to use MKL to do the matrix calculations
}
\examples{
x <- matrix(rnorm(1e4), 100)
y <- matrix(rnorm(1e2), 100)
XtX <- fMatProd(x, t(x))
invXtX <- fMatInv(XtX)
invXtXy <- fMatSolve(XtX, y)
}
