% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{flags}
\alias{flags}
\alias{LdFlags}
\alias{CxxFlags}
\title{Compilation flags for rmkl}
\usage{
LdFlags()

CxxFlags()
}
\description{
Output the compiler or linker flags required to build against `rmkl`.
These functions are typically called from `Makevars` as follows:
```
PKG_CXXFLAGS += $(shell "${R_HOME}/bin/Rscript" -e "rmkl::CxxFlags()")
PKG_LIBS += $(shell "${R_HOME}/bin/Rscript" -e "rmkl::LdFlags()")
```
\R packages using `rmkl` should also add the following to their `NAMESPACE` file:
```
importFrom(rmkl, CxxFlags)
importFrom(rmkl, LdFlags)
```
This is necessary to ensure that \pkg{rmkl} is loaded and available.
}
