% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.Regression.R
\name{prediction.Regression}
\alias{prediction.Regression}
\title{Prediction from MKL model}
\usage{
prediction.Regression(model, ktest, outcome)
}
\arguments{
\item{model}{MKL model}

\item{ktest}{Gramm matrix of training data and test data}

\item{outcome}{Outcome for the training data}
}
\value{
yhat Predicted value for each test point

predicted Sign of yhat, which is the final predicted outcome
}
\description{
This makes prediction for multiple kernel regression
}
