% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{rmkl.package.skeleton}
\alias{rmkl.package.skeleton}
\title{Create a skeleton for a new package depending on rmkl}
\usage{
rmkl.package.skeleton(name = "anRpackage", path = ".")
}
\arguments{
\item{name}{The name of your R package.}

\item{path}{The path of your R package.}
}
\value{
Nothing.
}
\description{
\code{rmkl.package.skeleton} automates the creation of a new source
package that intends to use features of `rmkl`.
It is based on the \link[pkgKitten]{kitten} function which it executes
first and append some options for \link[roxygen2]{roxygenize} function.
It will automatically generate a package document with needed `roxygen2`
parameters for a packaged based on `rmkl`.
}
\examples{
\dontrun{
rmkl.package.skeleton("testRMKLPkg")
}
}
\seealso{
\link[pkgKitten]{kitten}
}
