% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEMKL.Regression.R
\name{SEMKL.regression}
\alias{SEMKL.regression}
\title{Prediction from MKL model}
\usage{
SEMKL.regression(k, outcome, penalty, epsilon, tol = 1e-04, max.iters = 1000)
}
\arguments{
\item{k}{Gramm matrix of training data}

\item{outcome}{observed dependent variable}

\item{penalty}{Cost of unit miss fitted - observed}

\item{epsilon}{SVM parameter defining support vectors}

\item{tol}{Convergence criteria, algorithm stops once the biggest change of kernel weights
in two consectutive iterations is less than tol.}

\item{max.iters}{Termination criteria, algorithm will stop after 1000 iterations}
}
\value{
results Returns a list which includes model parameters, weights for kernels, and 
f which are the fitted values for the training set
}
\description{
This makes prediction for multiple kernel regression
}
