% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Surv_data.R
\docType{data}
\name{Surv_data}
\alias{Surv_data}
\title{Surv_data}
\format{
A dataframes with 4 columns and 1000 samples, where the x1 and x2
are generated from a multivariate distribution.
}
\usage{
Surv_data
}
\description{
Data from 1000 simulated patients with two variables with linear association
with the hazard function (h(x)=x1+2*x2), time, and vital status
}
\keyword{datasets}
