% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.small.R
\docType{data}
\name{tcga.small}
\alias{tcga.small}
\title{tcga.small}
\format{
A dataframe consisting a tiny subset of patients and genes from the TCGA
}
\usage{
tcga.small
}
\description{
A dataframe consisting a tiny subset of patients and genes from the TCGA
}
\keyword{datasets}
