% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{flags}
\alias{flags}
\alias{mklCxxFlags}
\alias{mklLdFlags}
\alias{rmklLibFlags}
\alias{rmklIncFlags}
\title{Compilation flags for rmkl}
\usage{
mklCxxFlags()

mklLdFlags()

rmklLibFlags()

rmklIncFlags()
}
\description{
Output the compiler or linker flags required to build against `rmkl`.
These functions are typically called from `Makevars` as follows:
```
PKG_CXXFLAGS += $(shell "${R_HOME}/bin/Rscript" -e "rmkl::rmklIncFlags()")
PKG_LIBS += $(shell "${R_HOME}/bin/Rscript" -e "rmkl::rmklLibFlags()")
```
\R packages using `rmkl` should also add the following to their `NAMESPACE` file:
```
importFrom(rmkl, rmklIncFlags)
importFrom(rmkl, rmklLibFlags)
```
This is necessary to ensure that \pkg{rmkl} is loaded and available.
}
