% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvMKL.R
\name{predict_Surv}
\alias{predict_Surv}
\title{Predict SurvMKL}
\usage{
predict_Surv(alpha, k0)
}
\arguments{
\item{alpha}{coefficient}

\item{k0}{the kernel cube needs prediction}
}
\value{
The predicted score
}
\description{
This function is used to predict SpicyMKL models
}
