% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvMKL.R
\name{predict_Surv}
\alias{predict_Surv}
\title{Predict SurvMKL}
\usage{
predict_Surv(alpha, k0)
}
\arguments{
\item{alpha}{coefficient}

\item{k0}{the kernel cube needs prediction}
}
\value{
The predicted score
}
\description{
This function is used to predict SpicyMKL models
}
\examples{
library(RMKL)
library(kernlab)
data(Surv_data)
rho0 <- .001*(Surv_data$status - seq(0, 10, length.out = dim(Surv_data)[1]))
klist <- list(kernelMatrix(rbfdot(2), as.matrix(Surv_data[,1:2])),
 	        kernelMatrix(vanilladot(), as.matrix(Surv_data[,1:2])))
kk <- simplify2array(klist)
modmkl <- SurvMKL(y = Surv_data$time, del = Surv_data$status, K = kk, rho = rho0, C = .005, lambda = .5, maxiter = 500, cri = .01)
predict_Surv(modmkl, kk)
}
