% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmkl-package.R
\docType{package}
\name{rmkl-package}
\alias{rmkl-package}
\title{`rmkl` Package}
\description{
The `rmkl` package aims to empower R with Intel MKL like Microsoft R Open (MRO)
by leveraging the Anaconda packages, `mkl`, `mkl-include` and `intel-openmp`.
However, it does not like MRO or `ropenblas` to directly affect R functions.
`rmkl` aims to act like a plug-in for R users and support dynamic linkage for Rcpp users.
`rmkl` supports windows/OS X/Linux. Besides, `rmkl` provides the ability to use Intel MKL in `Rcpp`,
so you can leverage MKL functions widely in `mkl_blas.h`, `mkl_cblas.h`, `mkl_spblas.h`,
`mkl_lapack.h` and `mkl_lapacke.h`.
}
